refactor: migrate all the pages to new version
