--- .backup/pages/_document.js
@@ -1,51 +1,0 @@
-import Document, { Html, Head, Main, NextScript } from "next/document";
-import { ServerStyleSheet } from "styled-components";
-
-class AppDocument extends Document {
-    static async getInitialProps(ctx) {
-        const sheet = new ServerStyleSheet();
-        const originalRenderPage = ctx.renderPage;
-
-        try {
-            ctx.renderPage = () => {
-                return originalRenderPage({
-                    enhanceApp: (App) => (props) => {
-                        return sheet.collectStyles(<App {...props} />)
-                    }
-                });
-            }
-
-            const initialProps = await Document.getInitialProps(ctx);
-
-            return {
-                ...initialProps,
-                styles: (
-                    <>
-                        {initialProps.styles}
-                        {sheet.getStyleElement()}
-                    </>
-                ),
-            }
-
-        } finally {
-            sheet.seal();
-        }
-    }
-
-    render() {
-        return (
-            <Html>
-                <Head>
-                    <link rel="preconnect" href="https://fonts.gstatic.com" />
-                    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:wght@400;600&display=swap" rel="stylesheet" />
-                </Head>
-                <body>
-                    <Main />
-                    <NextScript />
-                </body>
-            </Html>
-        )
-    }
-}
-
-export default AppDocument;
