+++ package.json
@@ -29,6 +29,7 @@
     "typescript": "4.9.5"
   },
   "devDependencies": {
+    "@types/react-custom-scrollbars": "^4.0.10",
     "autoprefixer": "^10.4.13",
     "postcss": "^8.4.21",
     "sass": "^1.58.2",
