+++ src/app/head.tsx
@@ -5,6 +5,9 @@ export default function Head() {
       <meta content="width=device-width, initial-scale=1" name="viewport" />
       <meta name="description" content="Generated by create next app" />
       <link rel="icon" href="/favicon.ico" />
+
+      {/* <link rel="preconnect" href="https://fonts.gstatic.com" />
+      <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:wght@400;600&display=swap" rel="stylesheet" /> */}
     </>
   )
 }
